Controller 2 support added and minor changes

+++ include/Bus.h
@@ -24,15 +24,15 @@ class Bus
 
         MIRROR getMirror();
 
-        void set_input(uint8_t state);
+        void set_input(uint16_t state);
         bool get_input();
     private:
         std::shared_ptr<PPU> ppu;
         std::shared_ptr<Cartridge> cart;
 
         bool NMI = false;
-        uint8_t controller_state;
-        uint8_t shift_register_controller;
+        uint16_t controller_state;
+        uint16_t shift_register_controller;
         bool handle_input;
         bool strobe;
